+++ src/ToolList.java
@@ -1,6 +1,6 @@
 public class ToolList {
     private final String[] toolList = {"Plow", "Watering can", "Fertilizer", "Pickaxe", "Shovel"};
-    private final int[] costList = {0,0, 10, 50, 7};
+    private final double[] costList = {0, 0, 10, 50, 7};
     private final double[] experienceList = {0.5, 0.5, 4, 15, 2};
 
     public int getIndexTool(String order) {
@@ -12,7 +12,7 @@ public class ToolList {
         return -1;
     }
 
-    public int costIndex(int index){
+    public double costIndex(int index){
         return costList[index];
     }
     public double expIndex(int index){
